backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
alembic==1.13.1
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pytesseract==0.3.10
pdf2image==1.17.0
opencv-python==4.9.0.80
numpy==1.26.3
Pillow==10.2.0
transformers>=4.47.0
torch>=2.2.0
torchvision>=0.17.0
python-dotenv==1.0.0
qwen-vl-utils
accelerate

